+++ tests/integration/test_old_timestamps.py
@@ -51,8 +51,7 @@ class TestContainersDir:
         old_ts_path = TMP_ROOT / OLD_TIMESTAMPS_NAME
         old_ts_path.touch()
         args = (PROGRAM_NAME, "-rtvv", str(TMP_ROOT))
-        res = cli.main(args)
-        assert res
+        cli.main(args)
         assert not old_ts_path.exists()
         self._assert_sizes(0)
 
@@ -64,8 +63,7 @@ class TestContainersDir:
         old_ts_path = child_root / OLD_TIMESTAMPS_NAME
         old_ts_path.touch()
         args = (PROGRAM_NAME, "-rtvv", str(TMP_ROOT))
-        res = cli.main(args)
-        assert res
+        cli.main(args)
         assert not old_ts_path.exists()
         self._assert_sizes(0, root=child_root)
 
@@ -77,7 +75,6 @@ class TestContainersDir:
         old_ts_path = TMP_ROOT / OLD_TIMESTAMPS_NAME
         old_ts_path.touch()
         args = (PROGRAM_NAME, "-rtvv", str(child_root))
-        res = cli.main(args)
-        assert res
+        cli.main(args)
         assert old_ts_path.exists()
         self._assert_sizes(0, root=child_root)