+++ tests/integration/test_one_container.py
@@ -33,8 +33,7 @@ class TestContainersDir:
     def test_containers_no_convert(self) -> None:
         """Test containers no convert."""
         args = (PROGRAM_NAME, "-x", "ZIP", str(TMP_ROOT / FN))
-        res = cli.main(args)
-        assert res
+        cli.main(args)
         for name, sizes in FNS.items():
             path = TMP_ROOT / name
             assert path.stat().st_size == sizes[1]