+++ picopt/handlers/webp.py
@@ -108,11 +108,11 @@ class WebPLossless(WebP):
         return new_path
 
 
-class WebPLossy(WebP):
-    """Handle lossy webp images."""
-
-    OUTPUT_FILE_FORMAT = FileFormat(WebP.OUTPUT_FORMAT_STR, False, False)
-    ARGS_PREFIX = (*WebP.ARGS_PREFIX, "-pass", "10", "-af")
+# class WebPLossy(WebP):
+#    """Handle lossy webp images."""
+#
+#    OUTPUT_FILE_FORMAT = FileFormat(WebP.OUTPUT_FORMAT_STR, False, False)
+#    ARGS_PREFIX = (*WebP.ARGS_PREFIX, "-pass", "10", "-af")
 
 
 class Gif2WebP(WebPBase):