+++ picopt/handlers/webp_animated.py
@@ -118,8 +118,8 @@ class WebPAnimatedLossless(WebPAnimatedBase):
     OUTPUT_FILE_FORMAT = FileFormat(WebPAnimatedBase.OUTPUT_FORMAT_STR, _LOSSLESS, True)
 
 
-class WebPAnimatedLossy(WebPAnimatedBase):
-    """Animated Lossy WebP Handler."""
-
-    _LOSSLESS = False
-    OUTPUT_FILE_FORMAT = FileFormat(WebPAnimatedBase.OUTPUT_FORMAT_STR, _LOSSLESS, True)
+# class WebPAnimatedLossy(WebPAnimatedBase):
+#    """Animated Lossy WebP Handler."""
+#
+#    _LOSSLESS = False
+#    OUTPUT_FILE_FORMAT = FileFormat(WebPAnimatedBase.OUTPUT_FORMAT_STR, _LOSSLESS, True)