+++ picopt.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 # Run picopt in development
-poetry run ./run.py "$@"
+poetry run ./picopt.py "$@"