+++ NEWS.md
@@ -1,10 +1,11 @@
 # 📰 Picopt News
 
-## v3.3.8
+## v3.4.0
 
 - Fix
   - Remove Lossy WebP optimization. Wasn't optimal.
   - Fix non zero exit code on success.
+  - `--destroy-metadata` option becomes `--strip-metadata`
 
 ## v3.3.7